requirements.txt
deep_sort_realtime==1.3.2
numpy==1.24.3
opencv_python==4.10.0.84
PyYAML==6.0
PyYAML==6.0.2
Requests==2.32.3
Shapely==2.0.6
torch==2.2.0
torchreid==0.2.5
twilio==9.3.7
